+++ lib/Gateway.ts
@@ -19,6 +19,7 @@ import ZwaveClient, {
 	ZUIValueId,
 	ZUIValueIdState,
 } from './ZwaveClient'
+import Cron from 'croner'
 
 import crypto from 'crypto'
 import { IMeterCCSpecific } from './Constants'
@@ -165,6 +166,14 @@ export type GatewayValue = {
 	retain?: boolean
 }
 
+export type ScheduledJob = {
+	name: string
+	cron?: string
+	enabled: boolean
+	runOnInit: boolean
+	code: string
+}
+
 export type GatewayConfig = {
 	type: GatewayType
 	payloadType?: PayloadType
@@ -182,6 +191,7 @@ export type GatewayConfig = {
 	logLevel?: LogLevel
 	logToFile?: boolean
 	values?: GatewayValue[]
+	jobs?: ScheduledJob[]
 	plugins?: string[]
 	logFileName?: string
 	manualDiscovery?: boolean
@@ -210,6 +220,7 @@ export default class Gateway {
 	private discovered: { [key: string]: HassDevice }
 	private topicLevels: number[]
 	private _closed: boolean
+	private jobs: Map<string, Cron> = new Map()
 
 	public get mqtt() {
 		return this._mqtt
@@ -267,6 +278,11 @@ export default class Gateway {
 
 			if (this.mqttEnabled) {
 				this._zwave.on('nodeStatus', this._onNodeStatus.bind(this))
+				this._zwave.on(
+					'nodeLastActive',
+					this._onNodeLastActive.bind(this)
+				)
+
 				this._zwave.on('valueChanged', this._onValueChanged.bind(this))
 				this._zwave.on('nodeRemoved', this._onNodeRemoved.bind(this))
 				this._zwave.on('notification', this._onNotification.bind(this))
@@ -281,6 +297,67 @@ export default class Gateway {
 			await this._zwave.connect()
 		} else {
 			logger.error('Z-Wave settings are not valid')
+		}
+	}
+
+	/**
+	 * Schedule a job
+	 */
+	scheduleJob(jobConfig: ScheduledJob) {
+		if (jobConfig.enabled) {
+			if (jobConfig.runOnInit) {
+				this.runJob(jobConfig).catch((error) => {
+					logger.error(
+						`Error while executing scheduled job "${jobConfig.name}": ${error.message}`
+					)
+				})
+			}
+
+			if (jobConfig.cron) {
+				try {
+					const job = new Cron(
+						jobConfig.cron,
+						this.runJob.bind(this, jobConfig)
+					)
+
+					if (job?.next()) {
+						this.jobs.set(jobConfig.name, job)
+						logger.info(
+							`Scheduled job "${jobConfig.name}" will run at ${job
+								.next()
+								.toISOString()}`
+						)
+					}
+				} catch (error) {
+					logger.error(
+						`Error while scheduling job "${jobConfig.name}": ${error.message}`
+					)
+				}
+			}
+		}
+	}
+
+	/**
+	 * Executes a scheduled job
+	 */
+	private async runJob(jobConfig: ScheduledJob) {
+		logger.info(`Executing scheduled job "${jobConfig.name}"...`)
+		try {
+			await this.zwave.driverFunction(jobConfig.code)
+		} catch (error) {
+			logger.error(
+				`Error executing scheduled job "${jobConfig.name}": ${error.message}`
+			)
+		}
+
+		const job = this.jobs.get(jobConfig.name)
+
+		if (job?.next()) {
+			logger.info(
+				`Next scheduled job "${jobConfig.name}" will run at ${job
+					.next()
+					.toISOString()}`
+			)
 		}
 	}
 
@@ -381,6 +458,18 @@ export default class Gateway {
 	}
 
 	/**
+	 * Method used to cancel all scheduled jobs
+	 */
+	cancelJobs() {
+		// cancel jobs
+		for (const [, job] of this.jobs) {
+			job.stop()
+		}
+
+		this.jobs.clear()
+	}
+
+	/**
 	 * Method used to close clients connection, use this before destroy
 	 */
 	async close(): Promise<void> {
@@ -391,6 +480,8 @@ export default class Gateway {
 		if (this._zwave) {
 			await this._zwave.close()
 		}
+
+		this.cancelJobs()
 
 		// close mqtt client after zwave connection is closed
 		if (this.mqttEnabled) {
@@ -1528,12 +1619,13 @@ export default class Gateway {
 						sensor.sensor,
 						sensor.objectId
 					)
-					Object.assign(cfg.discovery_payload, sensor.props || {})
 
 					// https://github.com/zwave-js/node-zwave-js/blob/master/packages/config/config/scales.json
 					if (valueId.unit) {
 						cfg.discovery_payload.unit_of_measurement = valueId.unit
 					}
+
+					Object.assign(cfg.discovery_payload, sensor.props || {})
 
 					// check if there is a custom value configuration for this valueID
 					if (valueConf) {
@@ -1978,11 +2070,48 @@ export default class Gateway {
 	}
 
 	/**
+	 * When a packet is received from a node to update it's last activity timestamp
+	 *
+	 */
+	private _onNodeLastActive(node: ZUINode): void {
+		if (!this.mqttEnabled) {
+			return
+		}
+
+		const nodeTopic = this.nodeTopic(node)
+
+		if (!this.config.ignoreStatus) {
+			let data: any
+
+			if (this.config.payloadType === PAYLOAD_TYPE.RAW) {
+				data = node.lastActive
+			} else {
+				data = {
+					time: Date.now(),
+					value: node.lastActive,
+				}
+			}
+
+			this._mqtt.publish(nodeTopic + '/lastActive', data)
+		}
+	}
+
+	/**
 	 * Driver status updates
 	 *
 	 */
 	private _onDriverStatus(ready: boolean): void {
 		logger.info(`Driver is ${ready ? 'READY' : 'CLOSED'}`)
+
+		this.cancelJobs()
+
+		if (ready) {
+			if (this.config.jobs?.length > 0) {
+				for (const jobConfig of this.config.jobs) {
+					this.scheduleJob(jobConfig)
+				}
+			}
+		}
 
 		this._mqtt.publish('driver/status', ready)
 	}