+++ src/views/ControlPanel.vue
@@ -140,12 +140,9 @@ export default {
 		StatisticsCard,
 	},
 	computed: {
-		...mapState(useBaseStore, ['nodes', 'zwave']),
+		...mapState(useBaseStore, ['nodes', 'zwave', 'controllerNode']),
 		timeoutMs() {
 			return this.zwave.commandsTimeout * 1000 + 800 // add small buffer
-		},
-		controllerNode() {
-			return this.nodes.find((n) => n.isControllerNode)
 		},
 		statisticsOpeningIndicator() {
 			return this.showControllerStatistics
@@ -271,6 +268,18 @@ export default {
 					icon: 'update',
 					desc: "Backup/Restore controller's NVM (Non Volatile Memory)",
 				},
+				{
+					text: 'Firmware update OTW',
+					options: [
+						{
+							name: 'Update',
+							action: 'firmwareUpdateOTW',
+						},
+					],
+					icon: 'update',
+					color: 'red',
+					desc: 'Perform a firmware update OTW (Over The Wire)',
+				},
 			],
 			rules: {
 				required: (value) => {
@@ -359,7 +368,7 @@ export default {
 						return
 					}
 				}
-				const args = []
+				let args = []
 				if (nodeId !== undefined) {
 					if (!broadcast) {
 						if (isNaN(nodeId)) {
@@ -402,6 +411,54 @@ export default {
 						}
 					)
 					if (!confirm || confirm !== 'yes') {
+						return
+					}
+				} else if (action === 'firmwareUpdateOTW') {
+					const result = await this.$listeners.showConfirm(
+						'Firmware update OTW',
+						`<h3 class="red--text">We don't take any responsibility if devices upgraded using Z-Wave JS don't work after an update. Always double-check that the correct update is about to be installed.</h3>
+						<h3 class="mt-2 red--text">A failure during this process may leave your controller in recovery mode, rendering it unusable until a correct firmware image is uploaded. In case of 500 series controllers a failure on this process is likely unrecoverable.</h3>
+						`,
+						'alert',
+						{
+							confirmText: 'Update',
+							width: 500,
+							inputs: [
+								{
+									type: 'file',
+									label: 'File',
+									hint: 'Firmware file',
+									key: 'file',
+								},
+							],
+						}
+					)
+
+					const file = result?.file
+
+					if (!file) {
+						return
+					}
+
+					try {
+						const buffer = await file.arrayBuffer()
+						args = [
+							{
+								name: file.name,
+								data: buffer,
+							},
+						]
+						const store = useBaseStore()
+
+						// start the progress bar
+						store.initNode({
+							id: this.controllerNode.id,
+							firmwareUpdate: {
+								progress: 0,
+							},
+						})
+					} catch (error) {
+						this.showSnackbar('Error reading file', 'error')
 						return
 					}
 				} else if (action === 'updateFirmware') {
@@ -488,13 +545,14 @@ export default {
 					}
 				} else if (action === 'driverFunction') {
 					const { data: snippets } = await ConfigApis.getSnippets()
-					const { code } = await this.$listeners.showConfirm(
+					await this.$listeners.showConfirm(
 						'Driver function',
 						'',
 						'info',
 						{
 							width: 900,
-							confirmText: 'Send',
+							confirmText: 'Close',
+							cancelText: '',
 							inputs: [
 								{
 									type: 'list',
@@ -517,10 +575,19 @@ export default {
 									},
 								},
 								{
+									type: 'button',
+									label: 'Run',
+									icon: 'play_circle_outline',
+									color: 'primary',
+									onChange: (values) => {
+										this.apiRequest(action, [values.code])
+									},
+								},
+								{
 									type: 'code',
 									key: 'code',
 									default:
-										'// Example:\n// const node = driver.controller.nodes.get(35);\n// await node.refreshInfo();',
+										'// Example:\n// const { logger, zwaveClient, require } = this\n// const node = driver.controller.nodes.get(35);\n// await node.refreshInfo();\n// logger.info(`Node ${node.id} is ready: ${node.ready}`);',
 									hint: `Write the function here. The only arg is:
                     <code>driver</code>. The function is <code>async</code>.`,
 								},
@@ -528,11 +595,7 @@ export default {
 						}
 					)
 
-					if (!code) {
-						return
-					}
-
-					args.push(code)
+					return
 				} else if (action === 'backupNVMRaw') {
 					const confirm = await this.$listeners.showConfirm(
 						'NVM Backup',
@@ -647,6 +710,10 @@ export default {
 					case 'restoreNVM':
 						this.showSnackbar('NVM restore DONE', 'success')
 						break
+					case 'firmwareUpdateOTW': {
+						// handled in App.vue
+						break
+					}
 					default:
 						this.showSnackbar(
 							'Successfully call api ' + data.api,
@@ -654,10 +721,28 @@ export default {
 						)
 				}
 			} else {
-				this.showSnackbar(
-					'Error while calling api ' + data.api + ': ' + data.message,
-					'error'
-				)
+				if (data.api === 'firmwareUpdateOTW') {
+					// this could happen when the update fails before start
+					// used to close the firmware update dialog
+					if (this.controllerNode.firmwareUpdate) {
+						useBaseStore().initNode({
+							id: this.controllerNode.id,
+							firmwareUpdate: false,
+							firmwareUpdateResult: {
+								success: false,
+								status: data.message,
+							},
+						})
+					}
+				} else {
+					this.showSnackbar(
+						'Error while calling api ' +
+							data.api +
+							': ' +
+							data.message,
+						'error'
+					)
+				}
 			}
 		},
 		bindEvent(eventName, handler) {